Imprve config passing
